Rearranges data to add navigation and GifCard
