+++ src/components/GifCards/GifCard/GifCard.css
@@ -1,0 +1,11 @@
+.GifCard {
+    z-index: 10px;
+    display: inline-block;
+}
+
+.GifCard img {
+    width: 400px;
+    height: 250px;
+    z-index: 10px;
+    display: inline-block;
+}
